outsourced interferometer to new module
