+++ doc/code/qml_templates.rst
@@ -1,7 +1,30 @@
 qml.templates
 =============
 
+This module provides a growing library of templates of common variational
+circuit architectures that can be used to easily build,
+evaluate, and train quantum nodes.
+
 .. currentmodule:: pennylane.templates
 
-.. automodule:: pennylane.templates
 
+Embeddings
+----------
+
+.. automodapi:: pennylane.templates.embeddings
+    :no-heading:
+    :include-all-objects:
+
+Layers
+------
+
+.. automodapi:: pennylane.templates.layers
+    :no-heading:
+    :include-all-objects:
+
+Subroutines
+-----------
+
+.. automodapi:: pennylane.templates.subroutines
+    :no-heading:
+    :include-all-objects:
