+++ doc/introduction/templates.rst
@@ -48,6 +48,22 @@ and the uniform parameter initialization strategy
 :func:`~.strong_ent_layers_uniform`.
 
 
+General templates
+-----------------
+
+Most generally, templates are just a collection of gates.
+
+The following general templates are available:
+
+:html:`<div class="summary-table">`
+
+.. autosummary::
+    :nosignatures:
+
+    ~pennylane.templates.Interferometer
+
+:html:`</div>`
+
 .. _intro_ref_temp_lay:
 
 Layer templates
@@ -55,9 +71,7 @@ Layer templates
 
 .. currentmodule:: pennylane.templates.layers
 
-Layer architectures, found in the :mod:`pennylane.templates.layers` module,
-define sequences of gates that are repeated like the layers in a neural network.
-They usually contain only trainable parameters.
+Layer architectures define sequences of trainable gates that are repeated like the layers in a neural network.
 
 The following layer templates are available:
 
@@ -68,7 +82,6 @@ The following layer templates are available:
 
     ~pennylane.templates.layers.CVNeuralNetLayer
     ~pennylane.templates.layers.CVNeuralNetLayers
-    ~pennylane.templates.layers.Interferometer
     ~pennylane.templates.layers.RandomLayer
     ~pennylane.templates.layers.RandomLayers
     ~pennylane.templates.layers.StronglyEntanglingLayer
@@ -83,10 +96,9 @@ The following layer templates are available:
 Embedding templates
 -------------------
 
-Embeddings, found in the :mod:`pennylane.templates.embeddings` module,
-encode input features into the quantum state of the circuit.
+Embeddings encode input features into the quantum state of the circuit.
 Hence, they take a feature vector as an argument. These embeddings can also depend on
-trainable parameters, in which case the embedding is learnable.
+trainable parameters and be repeated.
 
 The following embedding templates are available:
